add auth and contact ownership logic

+++ src/services/contacts.js
@@ -7,10 +7,12 @@ export const getAllContacts = async ({
   sortBy = '_id',
   sortOrder = 'asc',
   filters,
+  userId, //userId як параметр
 }) => {
   const skip = (page - 1) * perPage;
 
-  const query = {};
+  const query = { userId };
+
   if (filters.contactType) query.contactType = filters.contactType;
   if (typeof filters.isFavourite === 'boolean')
     query.isFavourite = filters.isFavourite;
@@ -37,14 +39,18 @@ export const createContact = async (contactData) => {
   return Contact.create(contactData);
 };
 
-export const updateContact = async (id, contactData) => {
-  const updatedContact = await Contact.findByIdAndUpdate(id, contactData, {
-    new: true,
-  });
+export const updateContact = async (id, contactData, userId) => {
+  const updatedContact = await Contact.findByIdAndUpdate(
+    { _id: id, userId },
+    contactData,
+    {
+      new: true,
+    },
+  );
   return updatedContact;
 };
 
-export const deleteContact = async (id) => {
-  const deletedContact = await Contact.findByIdAndDelete(id);
+export const deleteContact = async (id, userId) => {
+  const deletedContact = await Contact.findByIdAndDelete({ _id: id, userId });
   return deletedContact;
 };
